Avoid spurious drop-target indications
While app-view already prevented the default action for files and urls
in a browser window (navigation), the drag indicator showed that any
drop was possible. These changes cause the drag indicator to show that
nothing can be dropped on areas of the interface that don't contain
a drop target. The addition of dragenter avoids some flickering which
otherwise can occur when dragging a file rapidly across the group
or entry lists.

Accompanying changes to details-view, menu-item-view and open-view are
needed because those views relied on "inheriting" from app-view
the indication that dropping anything was allowed.