Adjust some overflow behavior in the details panel
Adding max-width, white-space, overflow and text-overflow to "a" children
of .details_field-value elements corrects an anomaly with long Website
values. Previously they displayed wider than other values, and the space
in which one could click to edit the field without triggering the link
was very thin. This establishes what was presumably the intended behavior.

Removing white-space from .details__attachment-preview-download-text
allows the text to wrap when the details pane is too narrow to show it all
on one line.

Replacing margin-right with max-width in .details__history fixes
issue #945 and makes fields in the history wrap as they do
in the ordinary details pane.

Adding flex to .details_history-top prevents the history details from
overlapping the navigation controls when the window height is too small
to show everything without scrolling.

Adding padding-right and margin-right to .details__history-top and
.details__history-buttons keeps those elements from overlapping
the scroll bar and aligns their right edges with the right edges
of the widest .details_field-value elements.